refactor initialization into separate red team policy class
